fixes to both sgl and vllm
