+++ functionary/prompt_template/prompt_template_v2.py
@@ -236,19 +236,25 @@ class PromptTemplateV2(PromptTemplate):
             add_all_recipient = True
             stage = "function"
 
-        return {
+        gen_state = {
             "stage": stage,
             "curr_tokens": curr_tokens,
             "curr_text": curr_text,
             "func_name": func_name,
             "func_index": -1,  # index of the tool in tool_calls
             "call_id": None,  # call_id of the current tool
-            "gen_empty_text": True,  # if first_time we return an empty delta with role=assistant
-            "first_time_func": True,
+            "first_chunk": True,
+            "first_function_chunk": True,
             "prev_newline": False,
             "add_all_recipient": add_all_recipient,
             "add_code_interpreter": add_code_interpreter,
         }
+
+        return (
+            self._update_gen_state_for_fn_call(gen_state, func_name)
+            if func_name is not None
+            else gen_state
+        )
 
     def stream_delta_text(
         self,
@@ -277,33 +283,50 @@ class PromptTemplateV2(PromptTemplate):
                 gen_state["prev_newline"] = True
             elif gen_state["prev_newline"] and delta_text != self.from_token:
                 responses.append(
-                    prompt_utils.get_text_delta_response("\n", True, finish_reason)
+                    prompt_utils.get_text_delta_response("\n", False, finish_reason)
                 )
                 gen_state["prev_newline"] = False
             elif gen_state["prev_newline"] is False:
-                if gen_state["gen_empty_text"]:
+                if gen_state["first_chunk"]:
                     responses.append(
                         prompt_utils.get_text_delta_response("", True, finish_reason)
                     )
-                    gen_state["gen_empty_text"] = False
+                    gen_state["first_chunk"] = False
+                    if gen_state["curr_text"] != "":
+                        responses.append(
+                            prompt_utils.get_text_delta_response(
+                                gen_state["curr_text"], False, finish_reason
+                            )
+                        )
                     delta_text = delta_text.lstrip(" ")
                 responses.append(
                     prompt_utils.get_text_delta_response(
                         delta_text, False, finish_reason
                     )
                 )
-        elif gen_state["stage"] == "parameter":
-            if gen_state["first_time_func"]:
+        elif gen_state["stage"] in ["parameter", "code-interpreter"]:
+            if gen_state["first_function_chunk"]:
                 responses.append(
                     prompt_utils.get_function_delta_response(
-                        gen_state, "", True, True, finish_reason
+                        gen_state, "", True, gen_state["first_chunk"], finish_reason
                     )
                 )
-                gen_state["first_time_func"] = False
+                gen_state["first_chunk"] = False
+                gen_state["first_function_chunk"] = False
+                if gen_state["curr_text"] != "":
+                    responses.append(
+                        prompt_utils.get_function_delta_response(
+                            gen_state,
+                            gen_state["curr_text"],
+                            False,
+                            False,
+                            finish_reason,
+                        )
+                    )
                 delta_text = delta_text.lstrip(" ")
             responses.append(
                 prompt_utils.get_function_delta_response(
-                    gen_state, delta_text, True, False, finish_reason
+                    gen_state, delta_text, False, False, finish_reason
                 )
             )
